[PM-22861] Account security state (#322)
## 🎟️ Tracking

https://bitwarden.atlassian.net/browse/PM-22861

## 📔 Objective

This adds the initial version of account security state to the sdk, on
init of the sdk state, and for enrollment to v2 crypto.

Account security state is an attestation to a specific version and
featureset that is enabled. The version is used to prevent cryptographic
downgrades of the account state.

This will prevent downgrades in the following ways:
- Arbitrary old states cannot be created by the server, because the
state is signed for a user by the user's signing key, the server can
only record entire states
- A signed in session will not accept a sync with an older state
- Some sign in mechanisms will not accept a sync with an older state
(login-with-device 2.0, PRF 2.0, both still to be specified)

A security version then is used to lock down specific features. If a
feature is to be disabled, then a migrator is needed. After migrating
the account not to use the retired feature, the version is incremented
and any subsequent syncs with that format included are rejected.

This PR stores the account security state on the client for v2 users.
However, besides validating that the state is correctly signed, it is
not yet used.

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes

---------

Co-authored-by: Daniel García <dani-garcia@users.noreply.github.com>