Create bitwarden-auth crate (#354)
## 🎟️ Tracking

<!-- Paste the link to the Jira or GitHub issue or otherwise describe /
point to where this change is coming from. -->

## 📔 Objective

Created a new auth crate for the work that auth is moving to the SDK,
and defined the WASM API for the send access functionality as a basic
skeleton.

Some things left open:
- We already have an `AuthClient` in `bitwarden-core`, it's not used for
WASM but it is for UniFFI. This PR leaves both clients in place but the
goal is to migrate the older clients functionality over to the auth
crate or somewhere else, and keep it out of core.
- Because we already have the `AuthClient` in core, I've made the
extension trait for auth use a function name like `auth_new`, this is to
avoid conflicts with the other function and should be renamed once the
older `AuthClient` is deleted.
- Currently `AuthClient` is wrapping a bitwarden-core `Client`, so it
can reuse the `ApiConfigurations` defined in core, this seems like it
would complicate the usage of the auth crate, so I'm thinking we should
extract the `ApiConfigurations` and maybe part of the token handling
code to a separate bitwarden-api crate that both core and auth can use
without depending on eachother?

## ⏰ Reminders before review

- Contributor guidelines followed
- All formatters and local linters executed and passed
- Written new unit and / or integration tests where applicable
- Protected functional changes with optionality (feature flags)
- Used internationalization (i18n) for all UI strings
- CI builds passed
- Communicated to DevOps any deployment requirements
- Updated any necessary documentation (Confluence, contributing docs) or
informed the documentation
  team

## 🦮 Reviewer guidelines

<!-- Suggested interactions but feel free to use (or not) as you desire!
-->

- 👍 (`:+1:`) or similar for great changes
- 📝 (`:memo:`) or ℹ️ (`:information_source:`) for notes or general info
- ❓ (`:question:`) for questions
- 🤔 (`:thinking:`) or 💭 (`:thought_balloon:`) for more open inquiry
that's not quite a confirmed
  issue and could potentially benefit from discussion
- 🎨 (`:art:`) for suggestions / improvements
- ❌ (`:x:`) or ⚠️ (`:warning:`) for more significant problems or
concerns needing attention
- 🌱 (`:seedling:`) or ♻️ (`:recycle:`) for future improvements or
indications of technical debt
- ⛏ (`:pick:`) for minor or nitpick changes