[PM-21691] - Added tree and test with implementation for CollectionView (#279)
## 🎟️ Tracking
[PM-21691](https://bitwarden.atlassian.net/browse/PM-21691)

## 📔 Objective
This splits the collections code out of the vault crate and adds a Tree
implementation for representing the parent child relationship between
nested collections.

This ideally will then be used as a unified representation for the
nested collection structure across all clients, instead of each one
having to implement their own.